Move method detect_os to mod os
